requirements.txt
Django
gunicorn
django-heroku
psycopg2
tweepy
requests
urllib3
iso8601
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
